requirements/dev.txt
-r base.txt
django-debug-toolbar==4.3.0
factory-boy==3.3.0
coverage==7.4.3
